fix: duplicate definitions of LibraryXBlockMetadata
