+++ openedx/core/djangoapps/content_libraries/api/__init__.py
@@ -1,6 +1,7 @@
 """
 Python API for working with content libraries
 """
+from .block_metadata import *
 from .collections import *
 from .containers import *
 from .courseware_import import *
