+++ openedx/core/djangoapps/content_libraries/api/blocks.py
@@ -55,6 +55,7 @@ from .exceptions import (
     InvalidNameError,
     LibraryBlockAlreadyExists,
 )
+from .block_metadata import LibraryXBlockMetadata, LibraryXBlockStaticFile
 from .containers import (
     create_container,
     get_container,
@@ -79,9 +80,6 @@ log = logging.getLogger(__name__)
 
 # The public API is only the following symbols:
 __all__ = [
-    # Models
-    "LibraryXBlockMetadata",
-    "LibraryXBlockStaticFile",
     # API methods
     "get_library_components",
     "get_library_block",
@@ -98,64 +96,6 @@ __all__ = [
     "delete_library_block_static_asset_file",
     "publish_component_changes",
 ]
-
-
-@dataclass(frozen=True, kw_only=True)
-class LibraryXBlockMetadata(PublishableItem):
-    """
-    Class that represents the metadata about an XBlock in a content library.
-    """
-    usage_key: LibraryUsageLocatorV2
-
-    @classmethod
-    def from_component(cls, library_key, component, associated_collections=None):
-        """
-        Construct a LibraryXBlockMetadata from a Component object.
-        """
-        last_publish_log = component.versioning.last_publish_log
-
-        published_by = None
-        if last_publish_log and last_publish_log.published_by:
-            published_by = last_publish_log.published_by.username
-
-        draft = component.versioning.draft
-        published = component.versioning.published
-        last_draft_created = draft.created if draft else None
-        last_draft_created_by = draft.publishable_entity_version.created_by if draft else None
-
-        return cls(
-            usage_key=library_component_usage_key(
-                library_key,
-                component,
-            ),
-            display_name=draft.title,
-            created=component.created,
-            modified=draft.created,
-            draft_version_num=draft.version_num,
-            published_version_num=published.version_num if published else None,
-            last_published=None if last_publish_log is None else last_publish_log.published_at,
-            published_by=published_by,
-            last_draft_created=last_draft_created,
-            last_draft_created_by=last_draft_created_by,
-            has_unpublished_changes=component.versioning.has_unpublished_changes,
-            collections=associated_collections or [],
-        )
-
-
-@dataclass(frozen=True)
-class LibraryXBlockStaticFile:
-    """
-    Class that represents a static file in a content library, associated with
-    a particular XBlock.
-    """
-    # File path e.g. "diagram.png"
-    # In some rare cases it might contain a folder part, e.g. "en/track1.srt"
-    path: str
-    # Publicly accessible URL where the file can be downloaded
-    url: str
-    # Size in bytes
-    size: int
-
 
 def get_library_components(
     library_key: LibraryLocatorV2,
